requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
motor>=3.3.0
beanie>=1.23.0,<2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pydantic[email]>=2.0.0
requests>=2.31.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0,<4.0.0
